+++ packages/riverpod/lib/src/notifier/family.dart
@@ -13,7 +13,7 @@ abstract class FamilyNotifier<State, Arg> extends BuildlessNotifier<State> {
     arg = element.origin.argument as Arg;
   }
 
-  /// {@macro riverpod.asyncnotifier.build}
+  /// {@macro riverpod.async_notifier.build}
   @visibleForOverriding
   State build(Arg arg);
 }
@@ -106,7 +106,7 @@ class NotifierProviderFamily<NotifierT extends FamilyNotifier<T, Arg>, T, Arg>
     required super.debugGetCreateSourceHash,
   }) : super(providerFactory: NotifierFamilyProvider.internal);
 
-  /// {@macro riverpod.overridewith}
+  /// {@macro riverpod.override_with}
   Override overrideWith(NotifierT Function() create) {
     return FamilyOverrideImpl<T, Arg,
         NotifierFamilyProvider<NotifierT, T, Arg>>(