+++ packages/riverpod/lib/src/state_provider/auto_dispose.dart
@@ -1,6 +1,6 @@
 part of '../state_provider.dart';
 
-/// {@macro riverpod.providerrefbase}
+/// {@macro riverpod.provider_ref_base}
 /// - [controller], the [StateController] currently exposed by this provider.
 abstract class AutoDisposeStateProviderRef<State>
     extends StateProviderRef<State> implements AutoDisposeRef<State> {}
@@ -48,7 +48,7 @@ class AutoDisposeStateProvider<T> extends _StateProviderBase<T> {
   @override
   late final Refreshable<StateController<T>> notifier = _notifier(this);
 
-  /// {@macro riverpod.overridewith}
+  /// {@macro riverpod.override_with}
   Override overrideWith(
     Create<T, AutoDisposeStateProviderRef<T>> create,
   ) {
@@ -91,7 +91,7 @@ class AutoDisposeStateProviderFamily<R, Arg> extends AutoDisposeFamilyBase<
               computeAllTransitiveDependencies(dependencies),
         );
 
-  /// {@macro riverpod.overridewith}
+  /// {@macro riverpod.override_with}
   Override overrideWith(
     R Function(AutoDisposeStateProviderRef<R> ref, Arg arg) create,
   ) {