+++ packages/riverpod/lib/src/stream_provider/auto_dispose.dart
@@ -1,14 +1,14 @@
 part of '../stream_provider.dart';
 
-/// {@macro riverpod.providerrefbase}
+/// {@macro riverpod.provider_ref_base}
 abstract class AutoDisposeStreamProviderRef<State>
     extends StreamProviderRef<State>
     implements AutoDisposeRef<AsyncValue<State>> {}
 
-/// {@macro riverpod.streamprovider}
+/// {@macro riverpod.stream_provider}
 class AutoDisposeStreamProvider<T> extends _StreamProviderBase<T>
     with AsyncSelector<T> {
-  /// {@macro riverpod.streamprovider}
+  /// {@macro riverpod.stream_provider}
   AutoDisposeStreamProvider(
     this._createFn, {
     super.name,
@@ -49,7 +49,7 @@ class AutoDisposeStreamProvider<T> extends _StreamProviderBase<T>
   @override
   late final Refreshable<Future<T>> future = _future(this);
 
-  /// {@macro riverpod.overridewith}
+  /// {@macro riverpod.override_with}
   Override overrideWith(
     Create<Stream<T>, AutoDisposeStreamProviderRef<T>> create,
   ) {
@@ -97,7 +97,7 @@ class AutoDisposeStreamProviderFamily<R, Arg> extends AutoDisposeFamilyBase<
           debugGetCreateSourceHash: null,
         );
 
-  /// {@macro riverpod.overridewith}
+  /// {@macro riverpod.override_with}
   Override overrideWith(
     Stream<R> Function(AutoDisposeStreamProviderRef<R> ref, Arg arg) create,
   ) {