+++ packages/riverpod_lint_flutter_test/test/lints/scoped_providers_should_specify_dependencies.dart
@@ -2,6 +2,7 @@
 
 import 'package:flutter/material.dart' as flutter;
 import 'package:flutter/material.dart' hide runApp;
+import 'package:flutter_test/flutter_test.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
@@ -119,6 +120,56 @@ void definitelyNotAMain() {
       child: Container(),
     ),
   );
+}
+
+void someTestFunction() {
+  final rootContainer = ProviderContainer(
+    overrides: [
+      scopedProvider.overrideWith((ref) => 0),
+      unimplementedScopedProvider.overrideWith((ref) => 0),
+      rootProvider.overrideWith((ref) => 0),
+    ],
+  );
+
+  testWidgets('override repositoryProvider in test', (tester) async {
+    await tester.pumpWidget(
+      ProviderScope(
+        overrides: [
+          scopedProvider.overrideWith((ref) => 0),
+          unimplementedScopedProvider.overrideWith((ref) => 0),
+          rootProvider.overrideWith((ref) => 0),
+        ],
+        child: Container(),
+      ),
+    );
+
+    await tester.pumpWidget(
+      ProviderScope(
+        parent: rootContainer,
+        overrides: [
+          scopedProvider.overrideWith((ref) => 0),
+          unimplementedScopedProvider.overrideWith((ref) => 0),
+          // expect_lint: scoped_providers_should_specify_dependencies
+          rootProvider.overrideWith((ref) => 0),
+        ],
+        child: Container(),
+      ),
+    );
+
+    await tester.pumpWidget(
+      Container(
+        child: ProviderScope(
+          overrides: [
+            scopedProvider.overrideWith((ref) => 0),
+            unimplementedScopedProvider.overrideWith((ref) => 0),
+            // expect_lint: scoped_providers_should_specify_dependencies
+            rootProvider.overrideWith((ref) => 0),
+          ],
+          child: Container(),
+        ),
+      ),
+    );
+  });
 }
 
 Widget fn() {