+++ packages/flutter_riverpod/CHANGELOG.md
@@ -6,6 +6,11 @@
 - **Breaking**: Removed everything marked as "deprecated"
 - Bumped minimum Dart SDK to >= 3.0.0-dev
 
+## Unreleased patch
+
+- Exceptions in asynchronous providers are now correctly received
+  by `ProviderObserver.providerDidFail`.
+
 ## 2.4.5 - 2023-10-28
 
 - Support assigning `AsyncValue<T>` to `AsyncNotifier<void>.state`