+++ packages/flutter_riverpod/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Exceptions in asynchronous providers are now correctly received
   by `ProviderObserver.providerDidFail`.
+- Fix exception when a `ProviderScope` is rebuilt with a different `key`.
 
 ## 2.4.5 - 2023-10-28